+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Geocoordinates reversed
+
 ## [0.6.4] - 2021-11-22
 
 ### Fixed